LPS-84119 Move method to BaseServiceObjectCheck, no logical changes
