+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/BaseServiceObjectCheck.java
@@ -1,0 +1,185 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.source.formatter.checks;
+
+import com.liferay.petra.string.StringPool;
+import com.liferay.portal.kernel.util.ArrayUtil;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.tools.ToolsUtil;
+import com.liferay.source.formatter.checks.util.SourceUtil;
+import com.liferay.source.formatter.util.FileUtil;
+
+import java.io.File;
+import java.io.IOException;
+
+import java.nio.file.FileVisitOption;
+import java.nio.file.FileVisitResult;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.SimpleFileVisitor;
+import java.nio.file.attribute.BasicFileAttributes;
+
+import java.util.ArrayList;
+import java.util.EnumSet;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import org.dom4j.Document;
+import org.dom4j.DocumentException;
+import org.dom4j.Element;
+
+/**
+ * @author Alan Huang
+ */
+public abstract class BaseServiceObjectCheck extends BaseJavaTermCheck {
+
+	@Override
+	public boolean isLiferaySourceCheck() {
+		return true;
+	}
+
+	@Override
+	protected String[] getCheckableJavaTermNames() {
+		return new String[] {JAVA_METHOD};
+	}
+
+	protected int getColumnIndex(
+		Element serviceXMLElement, String entityName, String columnName) {
+
+		for (Element entityElement :
+				(List<Element>)serviceXMLElement.elements("entity")) {
+
+			if (!entityName.equals(entityElement.attributeValue("name"))) {
+				continue;
+			}
+
+			int i = 0;
+
+			for (Element columnElement :
+					(List<Element>)entityElement.elements("column")) {
+
+				if (columnName.equals(columnElement.attributeValue("name"))) {
+					return i;
+				}
+
+				i++;
+			}
+		}
+
+		return -1;
+	}
+
+	protected String getPackageName(
+		String variableTypeName, List<String> importNames) {
+
+		for (String importName : importNames) {
+			if (importName.startsWith("com.liferay.") &&
+				importName.endsWith(".model." + variableTypeName)) {
+
+				return StringUtil.replaceLast(
+					importName, "." + variableTypeName, StringPool.BLANK);
+			}
+		}
+
+		return StringPool.BLANK;
+	}
+
+	protected Element getServiceXMLElement(String packageName) {
+		if (_serviceXMLElementsMap != null) {
+			return _serviceXMLElementsMap.get(packageName);
+		}
+
+		_serviceXMLElementsMap = new HashMap<>();
+
+		try {
+			_populateServiceXMLElements("modules/apps", 6);
+			_populateServiceXMLElements("modules/dxp/apps", 6);
+			_populateServiceXMLElements("portal-impl/src/com/liferay", 4);
+		}
+		catch (DocumentException | IOException exception) {
+			return null;
+		}
+
+		return _serviceXMLElementsMap.get(packageName);
+	}
+
+	private void _populateServiceXMLElements(String dirName, int maxDepth)
+		throws DocumentException, IOException {
+
+		File directory = getFile(dirName, ToolsUtil.PORTAL_MAX_DIR_LEVEL);
+
+		if (directory == null) {
+			return;
+		}
+
+		final List<File> serviceXMLFiles = new ArrayList<>();
+
+		Files.walkFileTree(
+			directory.toPath(), EnumSet.noneOf(FileVisitOption.class), maxDepth,
+			new SimpleFileVisitor<Path>() {
+
+				@Override
+				public FileVisitResult preVisitDirectory(
+					Path dirPath, BasicFileAttributes basicFileAttributes) {
+
+					String dirName = String.valueOf(dirPath.getFileName());
+
+					if (ArrayUtil.contains(_SKIP_DIR_NAMES, dirName)) {
+						return FileVisitResult.SKIP_SUBTREE;
+					}
+
+					Path path = dirPath.resolve("service.xml");
+
+					if (Files.exists(path)) {
+						serviceXMLFiles.add(path.toFile());
+
+						return FileVisitResult.SKIP_SUBTREE;
+					}
+
+					return FileVisitResult.CONTINUE;
+				}
+
+			});
+
+		for (File serviceXMLFile : serviceXMLFiles) {
+			Document serviceXMLDocument = SourceUtil.readXML(
+				FileUtil.read(serviceXMLFile));
+
+			Element serviceXMLElement = serviceXMLDocument.getRootElement();
+
+			String packagePath = serviceXMLElement.attributeValue(
+				"api-package-path");
+
+			if (packagePath == null) {
+				packagePath = serviceXMLElement.attributeValue("package-path");
+			}
+
+			if (packagePath != null) {
+				_serviceXMLElementsMap.put(
+					packagePath + ".model", serviceXMLElement);
+			}
+		}
+	}
+
+	private static final String[] _SKIP_DIR_NAMES = {
+		".git", ".gradle", ".idea", ".m2", ".settings", "bin", "build",
+		"classes", "dependencies", "node_modules", "node_modules_cache", "sql",
+		"src", "test", "test-classes", "test-coverage", "test-results", "tmp"
+	};
+
+	private Map<String, Element> _serviceXMLElementsMap;
+
+}
