+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/JavaServiceObjectCheck.java
@@ -14,47 +14,21 @@
 
 package com.liferay.source.formatter.checks;
 
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.TextFormatter;
-import com.liferay.portal.tools.ToolsUtil;
-import com.liferay.source.formatter.checks.util.SourceUtil;
 import com.liferay.source.formatter.parser.JavaTerm;
-import com.liferay.source.formatter.util.FileUtil;
 
-import java.io.File;
-import java.io.IOException;
-
-import java.nio.file.FileVisitOption;
-import java.nio.file.FileVisitResult;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.SimpleFileVisitor;
-import java.nio.file.attribute.BasicFileAttributes;
-
-import java.util.ArrayList;
-import java.util.EnumSet;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import org.dom4j.Document;
-import org.dom4j.DocumentException;
 import org.dom4j.Element;
 
 /**
  * @author Hugo Huijser
  */
-public class JavaServiceObjectCheck extends BaseJavaTermCheck {
-
-	@Override
-	public boolean isLiferaySourceCheck() {
-		return true;
-	}
+public class JavaServiceObjectCheck extends BaseServiceObjectCheck {
 
 	@Override
 	protected String doProcess(
@@ -67,16 +41,8 @@ public class JavaServiceObjectCheck extends BaseJavaTermCheck {
 			return javaTerm.getContent();
 		}
 
-		String javaTermContent = _formatGetterMethodCalls(
+		return _formatGetterMethodCalls(
 			javaTerm.getContent(), fileContent, importNames);
-
-		return _formatSetterMethodCalls(
-			javaTermContent, fileContent, importNames, absolutePath);
-	}
-
-	@Override
-	protected String[] getCheckableJavaTermNames() {
-		return new String[] {JAVA_METHOD};
 	}
 
 	private String _formatGetterMethodCalls(
@@ -108,160 +74,13 @@ public class JavaServiceObjectCheck extends BaseJavaTermCheck {
 		return content;
 	}
 
-	private String _formatSetterMethodCalls(
-		String content, String fileContent, List<String> importNames,
-		String absolutePath) {
-
-		Matcher matcher1 = _setterCallsPattern.matcher(content);
-
-		while (matcher1.find()) {
-			String setterCallsCodeBlock = matcher1.group();
-
-			String packageName = null;
-			String previousMatch = null;
-			String previousSetterObjectName = null;
-			String previousVariableName = null;
-			String variableTypeName = null;
-
-			Matcher matcher2 = _setterCallPattern.matcher(setterCallsCodeBlock);
-
-			List<String> allowedNonorderedSetterMethodTypes =
-				getAttributeValues(
-					_ALLOWED_NONORDERED_SETTER_METHOD_TYPES, absolutePath);
-
-			while (matcher2.find()) {
-				if (allowedNonorderedSetterMethodTypes.contains(
-						variableTypeName)) {
-
-					continue;
-				}
-
-				String match = matcher2.group();
-				String setterObjectName = TextFormatter.format(
-					matcher2.group(2), TextFormatter.I);
-				String variableName = matcher2.group(1);
-
-				if (!variableName.equals(previousVariableName)) {
-					previousMatch = match;
-					previousSetterObjectName = setterObjectName;
-					previousVariableName = variableName;
-
-					variableTypeName = getVariableTypeName(
-						content, fileContent, variableName);
-
-					packageName = _getPackageName(
-						variableTypeName, importNames);
-
-					continue;
-				}
-
-				Element serviceXMLElement = _getServiceXMLElement(packageName);
-
-				if (serviceXMLElement != null) {
-					int index1 = _getColumnIndex(
-						serviceXMLElement, variableTypeName,
-						previousSetterObjectName);
-					int index2 = _getColumnIndex(
-						serviceXMLElement, variableTypeName, setterObjectName);
-
-					if ((index2 != -1) && (index1 > index2)) {
-						int x = matcher2.start();
-
-						int y = content.lastIndexOf(previousMatch, x);
-
-						content = StringUtil.replaceFirst(
-							content, match, previousMatch, x);
-
-						return StringUtil.replaceFirst(
-							content, previousMatch, match, y);
-					}
-				}
-				else if (setterObjectName.compareTo(previousSetterObjectName) <
-							0) {
-
-					content = StringUtil.replaceFirst(
-						content, match, previousMatch, matcher2.start());
-
-					return StringUtil.replaceFirst(
-						content, previousMatch, match, matcher2.start());
-				}
-
-				previousMatch = match;
-				previousSetterObjectName = setterObjectName;
-				previousVariableName = variableName;
-			}
-		}
-
-		return content;
-	}
-
-	private int _getColumnIndex(
-		Element serviceXMLElement, String entityName, String columnName) {
-
-		for (Element entityElement :
-				(List<Element>)serviceXMLElement.elements("entity")) {
-
-			if (!entityName.equals(entityElement.attributeValue("name"))) {
-				continue;
-			}
-
-			int i = 0;
-
-			for (Element columnElement :
-					(List<Element>)entityElement.elements("column")) {
-
-				if (columnName.equals(columnElement.attributeValue("name"))) {
-					return i;
-				}
-
-				i++;
-			}
-		}
-
-		return -1;
-	}
-
-	private String _getPackageName(
-		String variableTypeName, List<String> importNames) {
-
-		for (String importName : importNames) {
-			if (importName.startsWith("com.liferay.") &&
-				importName.endsWith(".model." + variableTypeName)) {
-
-				return StringUtil.replaceLast(
-					importName, "." + variableTypeName, StringPool.BLANK);
-			}
-		}
-
-		return StringPool.BLANK;
-	}
-
-	private Element _getServiceXMLElement(String packageName) {
-		if (_serviceXMLElementsMap != null) {
-			return _serviceXMLElementsMap.get(packageName);
-		}
-
-		_serviceXMLElementsMap = new HashMap<>();
-
-		try {
-			_populateServiceXMLElements("modules/apps", 6);
-			_populateServiceXMLElements("modules/dxp/apps", 6);
-			_populateServiceXMLElements("portal-impl/src/com/liferay", 4);
-		}
-		catch (DocumentException | IOException exception) {
-			return null;
-		}
-
-		return _serviceXMLElementsMap.get(packageName);
-	}
-
 	private boolean _isBooleanColumn(
 		String variableTypeName, String getterObjectName,
 		List<String> importNames) {
 
-		String packageName = _getPackageName(variableTypeName, importNames);
+		String packageName = getPackageName(variableTypeName, importNames);
 
-		Element serviceXMLElement = _getServiceXMLElement(packageName);
+		Element serviceXMLElement = getServiceXMLElement(packageName);
 
 		if (serviceXMLElement == null) {
 			return false;
@@ -292,81 +111,7 @@ public class JavaServiceObjectCheck extends BaseJavaTermCheck {
 		return false;
 	}
 
-	private void _populateServiceXMLElements(String dirName, int maxDepth)
-		throws DocumentException, IOException {
-
-		File directory = getFile(dirName, ToolsUtil.PORTAL_MAX_DIR_LEVEL);
-
-		if (directory == null) {
-			return;
-		}
-
-		final List<File> serviceXMLFiles = new ArrayList<>();
-
-		Files.walkFileTree(
-			directory.toPath(), EnumSet.noneOf(FileVisitOption.class), maxDepth,
-			new SimpleFileVisitor<Path>() {
-
-				@Override
-				public FileVisitResult preVisitDirectory(
-					Path dirPath, BasicFileAttributes basicFileAttributes) {
-
-					String dirName = String.valueOf(dirPath.getFileName());
-
-					if (ArrayUtil.contains(_SKIP_DIR_NAMES, dirName)) {
-						return FileVisitResult.SKIP_SUBTREE;
-					}
-
-					Path path = dirPath.resolve("service.xml");
-
-					if (Files.exists(path)) {
-						serviceXMLFiles.add(path.toFile());
-
-						return FileVisitResult.SKIP_SUBTREE;
-					}
-
-					return FileVisitResult.CONTINUE;
-				}
-
-			});
-
-		for (File serviceXMLFile : serviceXMLFiles) {
-			Document serviceXMLDocument = SourceUtil.readXML(
-				FileUtil.read(serviceXMLFile));
-
-			Element serviceXMLElement = serviceXMLDocument.getRootElement();
-
-			String packagePath = serviceXMLElement.attributeValue(
-				"api-package-path");
-
-			if (packagePath == null) {
-				packagePath = serviceXMLElement.attributeValue("package-path");
-			}
-
-			if (packagePath != null) {
-				_serviceXMLElementsMap.put(
-					packagePath + ".model", serviceXMLElement);
-			}
-		}
-	}
-
-	private static final String _ALLOWED_NONORDERED_SETTER_METHOD_TYPES =
-		"allowedNonorderedSetterMethodTypes";
-
-	private static final String[] _SKIP_DIR_NAMES = {
-		".git", ".gradle", ".idea", ".m2", ".settings", "bin", "build",
-		"classes", "dependencies", "node_modules", "node_modules_cache", "sql",
-		"src", "test", "test-classes", "test-coverage", "test-results", "tmp"
-	};
-
 	private static final Pattern _getterCallPattern = Pattern.compile(
 		"\\W(\\w+)\\.\\s*(get)([A-Z]\\w*)\\(\\)");
-	private static final Pattern _setterCallPattern = Pattern.compile(
-		"(\\w+)\\.\\s*set([A-Z]\\w*)\\([^;]+;");
-	private static final Pattern _setterCallsPattern = Pattern.compile(
-		"(^[ \t]*\\w+\\.\\s*set[A-Z]\\w*\\([^;]+;\n)+",
-		Pattern.DOTALL | Pattern.MULTILINE);
-
-	private Map<String, Element> _serviceXMLElementsMap;
 
 }
